+++ fiddles/node/fiddle-0021-DirectoryWatcher/app.js
@@ -1,0 +1,43 @@
+#!/usr/bin/env node --harmony
+
+"use strict";
+
+var watchman = require('fb-watchman'),
+  client = new watchman.Client(),
+  dir_of_interest = "../../../fiddles";
+
+client.capabilityCheck({optional:[], required:['relative_root']},
+  function (error, resp) {
+    if (error) {
+      console.log(error);
+      client.end();
+      return;
+    }
+
+    // Initiate the watch
+    client.command(['watch-project', dir_of_interest],
+      function (error, resp) {
+        if (error) {
+          console.error('Error initiating watch:', error);
+          return;
+        }
+
+        // It is considered to be best practice to show any 'warning' or
+        // 'error' information to the user, as it may suggest steps
+        // for remediation
+        if ('warning' in resp) {
+          console.log('warning: ', resp.warning);
+        }
+
+        // `watch-project` can consolidate the watch for your
+        // dir_of_interest with another watch at a higher level in the
+        // tree, so it is very important to record the `relative_path`
+        // returned in resp
+
+        console.log('watch established on ', resp.watch,
+          ' relative_path', resp.relative_path);
+      });
+  });
+
+
+