+++ CHANGELOG.md
@@ -1269,4 +1269,8 @@ Alternately, the _Daily Diary of my Programming Dreams_.
         9.  Updated [scripts/fiddle-start.sh](scripts/fiddle-start.sh)
         10. Updated [fiddles/android/README.md](fiddles/android/README.md)
 
- 
+* **Java**
+  1.  Started [fiddles/java/fiddle-0004-CSVToJsonEndpoint](fiddles/java/fiddle-0004-CSVToJsonEndpoint)
+
+* **Node**
+  1.  Started [fiddles/node/fiddle-0021-DirectoryWatcher](fiddles/node/fiddle-0021-DirectoryWatcher)